bucketmaniac/requirements.txt
pkg-resources==0.0.0
pytaglib==1.2.0
python-vlc==1.1.2
urwid==1.3.1
